+++ examples/get-started/main.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     elestio = {
       source  = "elestio/elestio"
-      version = "1.0.0" # check out the latest version in the release section
+      version = "0.2.0" # check out the latest version in the release section
     }
   }
 }