+++ examples/index/env.tf
@@ -1,6 +1,0 @@
-# You can omit the provider block
-# provider "elestio" {}
-
-resource "elestio_service" "myservice" {
-  # ...
-}