+++ examples/index/static.tf
@@ -1,8 +1,0 @@
-provider "elestio" {
-  email     = "YOUR-EMAIL"
-  api_token = "YOUR-API-TOKEN"
-}
-
-resource "elestio_project" "myproject" {
-  # ...
-}